change type balance

+++ src/main/java/com/boa/web/response/cardsrequest/Card.java
@@ -24,7 +24,7 @@ public class Card {
     @JsonProperty("currency")
     private String currency;
     @JsonProperty("available-balance")
-    private Integer availableBalance;
+    private Double availableBalance;
     @JsonProperty("type")
     private Type type;
     @JsonProperty("category")
@@ -99,12 +99,12 @@ public class Card {
     }
 
     @JsonProperty("available-balance")
-    public Integer getAvailableBalance() {
+    public Double getAvailableBalance() {
         return availableBalance;
     }
 
     @JsonProperty("available-balance")
-    public void setAvailableBalance(Integer availableBalance) {
+    public void setAvailableBalance(Double availableBalance) {
         this.availableBalance = availableBalance;
     }
